♻️: Post 타입, ApiResponseWith 타입 분리 및 계층 변경
- ApiResponseWith:
  - api 응답 타입인 ResponseWithData 타입을 ApiResponseWith로 함수명 변경
  - 계층을 shared/api로 변경
- Post, Posts 타입 계층 변경
  - 도메인 엔티티인 Post는 entities의 posts/model 계층으로, api 응답 결과인 Posts는 추후 서버 상태 관리를 통한 응답값으로 사용되므로 features의 posts/api 계층으로 이동
