+++ src/entities/posts/model/index.ts
@@ -1,0 +1,1 @@
+export type { Post } from './Post';
