+++ src/features/posts/model/Posts.ts
@@ -1,0 +1,4 @@
+import type { Post } from '@/entities/posts/model';
+import type { ApiResponseWith } from '@/shared/api';
+
+export type Posts = ApiResponseWith<Post, 'posts'>;
